% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{mymean}
\alias{mymean}
\title{Function to Calculate Mean}
\usage{
mymean(x, na.rm = F)
}
\arguments{
\item{x}{A numerical vector}

\item{na.rm}{= TRUE}
}
\value{
This function returns the mean of a numerical vector
}
\description{
Some description of this function
}
\examples{
x <- c(2, 4, 6, 8)
mymean(x)
}
