% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_root.R
\name{sign_root}
\alias{sign_root}
\title{Take the nth root of a value, but keep the sign intact}
\usage{
sign_root(n, value)
}
\arguments{
\item{n}{root to take(n= 2 would correspond to the square root)}

\item{value}{value to take the root of}
}
\value{
the nth root of abs(value) with the sign of value
}
\description{
Take the nth root of a value, but keep the sign intact
}
